[Refactor] #112 미리 정의한 Throwable들을 Domain 모듈로 이동
